documentation: apply dual license to the OP-TEE header definitions
The licensing terms for the OP-TEE OS documentation is BSD-2-Clause, as
mentioned in the LICENSE file. This commit applies a dual license
(BSD-2-Clause OR GPL-2.0) to the OP-TEE header definitions (struct
optee_header, struct optee_header_v2, struct optee_image) mentioned in
optee_design.md. The intention is to make it easier for GPL-licensed
projects (bootloaders, typically) to interface with OP-TEE. One example
of GPL-2.0 bootloader is Barebox [1].

Link: [1] https://www.barebox.org/
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Suggested-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>