core: PTAs get core session_id to identify sessions
Prior this change, PTA handled an opaque session references as
session client identifier as stated in the GPD TEE Client API
and Internal Core API specifications. There was a specific field
usr_ctx in the generic TEE session structure (struct tee_ta_session)
to save this reference generated a session opening and use it
on later session invocation.

This change removes this @usr_ctx field and changes the PTA
interface so that PTA now rely in the unique session ID produced
by OP-TEE core to distinguish between client sessions. As a
consequence, PTA open_session handle no more generate the session
context handle but uses the input session ID. It is up to PTAs
to use this information of not.

This change updates pseudo TA driver and the embedded PTAs according
to this API change.

Note that the 2 only PTAs that generated such session reference are
pta_invoke_test.c and pta_socket. Both are updated to use the 32bit
session ID instead of void * references.

This change also replace some camelcase labels into lowercase
labels.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>