[RFC] Add support for runtime dynamic linking in TAs (libdl)
This is a proof-of-concept implementing some support for runtime
dynamic linking in Trusted Applications. A new library called libdl
is provided, with the following functions: dlopen(), dlsym() and
dlclose(). They are similar to the POSIX functions of the same name.

Since the ELF loader for TAs is in kernel space, this commit also adds
two "system calls", which are actually implemented in the system
pseudo-TA (core/arch/arm/pta/system.c). The libdl library simply
invokes the PTA.

Tested with xtest 1021 (see my other RFC patch in optee_test) on QEMU
and HiKey960.

Main limitations/known bugs:
  - Only (RTLD_NOW | RTLD_GLOBAL) is supported
  - Libraries are NOT unmapped when the last reference is dropped
    (actually, library handles are not even reference counted)
  - Some code duplication and other things to improve (see FIXME)
  - There is a bug with 64-bit TAs on HiKey960 or QEMUv8 which causes a
    "prefetch abort" in user space when the TA access the address
    returned by dlsym(), although it looks fine and certainly is in the
    range of the executable section of the library (which is reportedly
    mapped with the proper permissions according to the crash dump). It
    looks like TTBR0 contains an invalid value. This needs further
    investigation.
  - Does not work with CFG_ULIBS_SHARED=y introduced recently in
    commit 64718c936163 ("Allow building libutils etc. as shared
    libraries"). The symptom is: dlopen() fails with
    TEE_ERROR_ACCESS_CONFLICT in umap_add_region().

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>