libutils: ctype.h: define functions as builtins only when supported
Our <ctype.h> defines a number of functions as macros that evaluate to
the builtin function of the same name, for instance:

 #define isalpha(__c) __builtin_isalpha(__c)

The goal is to allow the compiler to recognize the functions as
builtins, therefore enabling some optimizations.

In preparation for supporting Clang, which does not implement all the
builtins that GCC recognizes, we need to test builtin support more
precisely.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>