Error out if CFG_ULIBS_GPROF and CFG_ULIBS_SHARED are both enabled
When CFG_ULIBS_GPROF=y, the TA linker script references the
__utee_mcount symbol, which is external to the TA if libutee is built
as a shared library. This causes the following link error:

 ta.lds:19: unresolvable symbol `__utee_mcount' referenced in expression

It should quite easy to get rid of the allocation of the profiling
buffer in the TA's .bss segment, once we have a syscall (or a PTA
command) to allocate and map TA memory from the TEE core. Until then,
let's just prevent the incompatible configuration. Profiling is just a
debugging tool, so it should not be a problem to force static linking in
that case.

Fixes: 64718c936163 ("Allow building libutils etc. as shared libraries")
Reported-by: Sumit Garg <sumit.garg@linaro.org>
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>