stmp32mp1: support STPMIC1 device
Implement a power management device driver for the platform to handle
STPMIC1 as external regulator control device. This driver is embedded upon
CFG_STPMIC1=y and get the regulator definition and configuration from the
device tree.

Platform implements stm32mp_with_pmic() which returns whether or not a
PMIC device is used to control external regulators.

Signed-off-by: Etienne Carriere <etienne.carriere@st.com>