core: arm: kern.ld.S: ignore .init section
The ELF .init section is meant to be used by program loaders to run
special initialization code before the main entry point is called. This
does not apply to the TEE core, and the compiler does not generate such
a section. Therefore, mentioning it in the linker script is useless.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>