[Review] core: arm32: add option for PIE core
Fix the assembler relocation code so that is does not contain a
relocation itself. In theory, this should be enough to make it possible
to link tee.elf at any address, but it looks like there are some wrong
assumptions in the code because if I link tee.elf at address zero
(". = 0;" instead of ". = TEE_TEXT_VA_START" in kern.ld.S, and write
the value of TEE_TEXT_VA_START in tee-init_load_addr.txt so that the
header passed to the bootloader is unchanged), the boot panics with:

 E/TC:0 0 Panic 'Impossible memory alignment' at core/arch/arm/mm/core_mmu.c:894 <init_mem_map>

Another issue is the pager. I think the init code needs to be moved
before the relocations are applied. This would mean re-writing this
code, too, to make it is position-independent.

Signed-off-by: Jerome Forissier <jerome@forissier.org>